Add a new folder with task.js module
