+++ src/index.html
@@ -45,8 +45,7 @@
         Clear all completed
       </button>
     </div>
-    <script type="module" src="./modules/tasks.js"></script>
-    <script type="module" src="./modules/task-item.js"></script>
     <script type="module" src="index.js"></script>
+    <script type="module" src="modules/tasks.js"></script>
   </body>
 </html>
